attempt at adding --force option to svk rm
we have a --force flag to "svk rm" that works for simple files

test that --force works for deleting directories with modified files in too

really test deleting the directory, not the modified file again

add tests for 'svk rm --force A/' when A/quux is a file not in
the repository. Updated the code so that it will be deleted but
not scheduled for deletion; this is necessary to avoid confusing
'svk revert'. 

when trying to delete files or directories with modifications,
hint about the '--force' flag instead of telling them to 'svk
revert' the file/dir first.

update expected error message for symlink test too

when deleting a directory with file scheduled for addition in it,
don't schedule it for deletion (just unschedule its add)

when doing 'svk delete --force foo' and foo is a file scheduled
for addition, unschedule the file (currently we remove it too,
but that might change).

minor tidy to SVK::XD->do_delete

pass along full copaths, not the reportpaths.
Also get rid of regex-matching grep by using hash lookup instead.

report paths correctly

use correct report root even if report is empty

minor refactoring; do not check same variable many times

minor cleanup: put code closer to where it's used

added comment

when something is scheduled "with history", don't unschedule it

* ignore files that we are not meant to see
* print "D foo" if foo was added, then immediately deleted

improve diagnostic message

git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/trunk@1584 4ccb8a5b-8ae0-0310-9b02-86531366991d