Lump commit
Fix a bug that 'svk update' can assume part of your directory are
already up-to-date even when they are not, hence update the meta
data directly without updating the file.  When there are changes
in-between, your checkout can contain the reverse modification
of the latest content in the repository.

This has been bugging us for quite a while.  The code causing
the bug was found while refactoring on the view branch.


forgot to commit tests for r1639.

git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/branches/1.0-releng@1642 4ccb8a5b-8ae0-0310-9b02-86531366991d