Date: Thu, 23 Mar 2006 15:32:49 -0700 To: svk-dev@lists.openfoundry.org From: Michael Hendricks <michael@ndrix.org>
I've attached a patch which removes SVK's dependency on the TimeDate
CPAN module.  Essentially, it steals the important guts from TimeDate
and makes use of POSIX::strftime for date formatting.

As an additional bonus, the log command now runs about 25% faster when
generating long logs.

This is the second patch in a series with the theme: "Reduce SVK's
dependencies where it makes sense".  I sent the first patch to the list
some time ago and it removed the dependency on Pod::Simple.  This patch
does not depend on that one.


Log:
 r1763@askos:  michael | 2006-03-17 20:43:03 -0700
 Creating a local branch for removing the dependency on Date::Parse.
 r1764@askos:  michael | 2006-03-17 20:44:29 -0700
 Replaced one occurence of str2time with a custom str2time based on the raw Date::Parse::str2time code with optimizations for the svn:date case.
 r1765@askos:  michael | 2006-03-17 21:31:41 -0700
 Make use of the new str2time replacement subroutine
 r1766@askos:  michael | 2006-03-23 14:59:15 -0700
 Created a new my_time2str sub to replace the one from Date::Format
 r1767@askos:  michael | 2006-03-23 15:05:35 -0700
 Use the new my_time2str sub to replate time2str in SVK::Command::Log.
 r1768@askos:  michael | 2006-03-23 15:19:25 -0700
 Cleaned up the helper subs a little bit

git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/trunk@1763 4ccb8a5b-8ae0-0310-9b02-86531366991d