Added TODO test for repeated smerges from tags of trunk to a branch.
Prevent unwanted repeated merges when repeatedly merging tags to a branch.
We now also record the copy source(s) of the path being merged from in the
merge ticket.


Fixed so that the test succeeds 100% of the time instead of 50% of the
time.  It was failing because, although merge tickets are sorted,
repository uuids are random and therefore merge tickets from one
repository would randomly end up either before of after the ones from
the other repository.


add t/07smerge-tags.t to MANIFEST.

git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/branches/2.0-releng@2337 4ccb8a5b-8ae0-0310-9b02-86531366991d